# docs/swagger.yaml
components:
  schemas:
...
    github_com_array_banking-api_internal_dto.CreateAccountRequest:
      properties:
        accountType:
          enum:
          - checking
          - savings
          - money_market
          type: string
        initialDeposit:
          type: string
      required:
      - accountType
      type: object
    github_com_array_banking-api_internal_dto.CreateAccountResponse:
      properties:
        account:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_models.Account'
        message:
          type: string
      type: object
...
    github_com_array_banking-api_internal_dto.UpdateAccountStatusRequest:
      properties:
        status:
          enum:
          - active
          - inactive
          - frozen
          - closed
          type: string
      required:
      - status
      type: object
...
    github_com_array_banking-api_internal_errors.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorDetail'
      type: object
    github_com_array_banking-api_internal_models.Account:
      properties:
        account_number:
          type: string
        account_type:
          type: string
        balance:
          type: number
        closed_at:
          type: string
        created_at:
          type: string
        currency:
          type: string
        deleted_at:
          $ref: '#/components/schemas/gorm.DeletedAt'
        id:
          type: string
        interest_rate:
          type: number
        status:
          type: string
        updated_at:
          type: string
        user_id:
          type: string
      type: object
...
    gorm.DeletedAt:
      properties:
        time:
          type: string
        valid:
          description: Valid is true if Time is not NULL
          type: boolean
      type: object
...
    message:
      properties:
        message:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
        eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: support@arraybanking.com
    name: API Support
  description: Production-quality banking REST API for developer assessment and interviewing.
    Provides core banking functionality including identity management, account operations,
    customer management, transaction processing, NorthWind Bank integration, and regulator
    compliance notifications.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Array Banking API
  version: "1.0"
openapi: 3.1.0
paths:
  /accounts:
    get:
      description: Retrieve all bank accounts belonging to the authenticated user
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/github_com_array_banking-api_internal_models.Account'
                type: array
          description: List of user's accounts
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get all user accounts
      tags:
      - Accounts
    post:
      description: Create a new bank account (checking, savings, or money_market)
        with optional initial deposit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.CreateAccountRequest'
        description: Account creation details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.CreateAccountResponse'
          description: Account created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: TRANSACTION_002 - Invalid initial deposit amount
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - Accounts
  /accounts/{accountId}:
    delete:
      description: Permanently close an account. Account must have zero balance to
        be closed.
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/message'
                properties:
                  data:
                    type: object
                  message:
                    type: string
                  meta:
                    type: object
                type: object
          description: Account closed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_003 - Invalid account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_005 - Account has non-zero balance
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Close account
      tags:
      - Accounts
    get:
      description: Retrieve detailed information about a specific account belonging
        to the authenticated user
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_models.Account'
          description: Account details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_003 - Invalid account ID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - Accounts
  /accounts/{accountId}/status:
    patch:
      description: Update the status of an account (active, inactive, frozen, closed)
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.UpdateAccountStatusRequest'
        description: New account status
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_models.Account'
          description: Updated account details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid request body or account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Update account status
      tags:
      - Accounts
...
servers:
- description: Local development server
  url: http://localhost:8080
- description: Production server
  url: https://api.arraybanking.com